refactor(git-routes): extract chat-scoped handlers, fix diff range + Zod
Three issues from code review addressed in one move:

HIGH: handleDiffSinceMain compared mergeBase..HEAD, so uncommitted
working-tree edits were invisible in the Review Panel even though the
panel is meant to show the chat's working changes. Switch to
'git diff --name-status <mergeBase>' (no ..HEAD) so the result includes
both committed branch changes AND uncommitted edits. Untracked files are
still excluded.

MEDIUM: handleDiffSinceMain read req.body.chatId / req.body.files
directly. Add a DiffSinceMainBody Zod schema; chatId is optional (the
:id route param can substitute via getEffectivePath), files is an
optional string array.

LOW: git.ts grew to 513 lines. Pull the chat-scoped routes out into a
new git-chat.ts (POST /api/git/{status,stage,unstage,commit,push} and
POST /api/projects/:id/git/diff-since-main). git.ts is now 210 lines,
git-chat.ts is 254 — both under the 300-line limit.

git-chat introduces a small chatRoute<T> helper that handles parse +
chatId resolve + optional path validation + standard error mapping, so
the per-handler code stays one to three lines.
