+++ packages/core/src/server/routes/git.ts
@@ -1,6 +1,5 @@
 import { Router, Request, Response } from 'express';
 import { readFile } from 'node:fs/promises';
-import { z } from 'zod';
 import type { RouteContext } from './types.js';
 import { getEffectivePath, param } from './types.js';
 import { resolveAndValidatePath } from './path-utils.js';
@@ -8,6 +7,7 @@ import { asyncHandler } from './async-handler.js';
 import { GitService } from '../../git/git-service.js';
 import { createChildLogger } from '../../logger.js';
 import { gitWriteRoutes } from './git-write.js';
+import { gitChatRoutes } from './git-chat.js';
 
 const logger = createChildLogger('routes:git');
 
@@ -183,309 +183,9 @@ async function handleDiff(ctx: RouteContext, req: Request, res: Response): Promi
   }
 }
 
-const StatusRequestSchema = z.object({
-  chatId: z.string(),
-});
-
-const PushRequestSchema = z.object({
-  chatId: z.string(),
-});
-
-const StageRequestSchema = z.object({
-  chatId: z.string(),
-  files: z.array(z.string()),
-});
-
-const CommitRequestSchema = z.object({
-  chatId: z.string(),
-  message: z.string().min(1, 'Commit message cannot be empty'),
-  files: z.array(z.string()),
-});
-
-function parsePortcelainStatus(output: string): { staged: string[]; unstaged: string[]; untracked: string[] } {
-  const staged: string[] = [];
-  const unstaged: string[] = [];
-  const untracked: string[] = [];
-
-  for (const line of output.split('\n').filter(Boolean)) {
-    const indexStatus = line[0] ?? ' ';
-    const workingStatus = line[1] ?? ' ';
-    const filename = line.slice(3);
-
-    if (indexStatus === '?' && workingStatus === '?') {
-      untracked.push(filename);
-      continue;
-    }
-    if (indexStatus !== ' ') staged.push(filename);
-    if (workingStatus !== ' ') unstaged.push(filename);
-  }
-
-  return { staged, unstaged, untracked };
-}
-
-/** POST /api/git/push */
-async function handleGitPush(ctx: RouteContext, req: Request, res: Response): Promise<void> {
-  const parsed = PushRequestSchema.safeParse(req.body);
-  if (!parsed.success) {
-    res.status(400).json({ error: 'chatId is required' });
-    return;
-  }
-
-  const { chatId } = parsed.data;
-  const workDir = ctx.chats.getEffectivePath(chatId);
-  if (!workDir) {
-    res.status(404).json({ error: 'Chat not found' });
-    return;
-  }
-
-  try {
-    const svc = GitService.forProject(workDir);
-    const result = await svc.push();
-    if (result.status === 'rejected') {
-      res.status(400).json({ error: result.message ?? 'Push rejected' });
-      return;
-    }
-    res.json({ success: true });
-  } catch (err) {
-    logger.error({ err, workDir, chatId }, 'Failed to push');
-    res.status(400).json({ error: (err as Error).message ?? 'Unknown error' });
-  }
-}
-
-/** POST /api/git/commit */
-async function handleGitCommit(ctx: RouteContext, req: Request, res: Response): Promise<void> {
-  const parsed = CommitRequestSchema.safeParse(req.body);
-  if (!parsed.success) {
-    const firstIssue = parsed.error.issues[0];
-    res.status(400).json({ error: firstIssue?.message ?? 'Invalid request' });
-    return;
-  }
-
-  const { chatId, message, files } = parsed.data;
-  const workDir = ctx.chats.getEffectivePath(chatId);
-  if (!workDir) {
-    res.status(404).json({ error: 'Chat not found' });
-    return;
-  }
-
-  for (const file of files) {
-    const resolved = resolveAndValidatePath(workDir, file);
-    if (!resolved) {
-      res.status(400).json({ error: `Path outside project: ${file}` });
-      return;
-    }
-  }
-
-  try {
-    const svc = GitService.forProject(workDir);
-    if (files.length > 0) {
-      await svc.stage(files);
-    }
-    const hash = await svc.commit(message);
-    res.json({ hash });
-  } catch (err) {
-    logger.error({ err, workDir, chatId }, 'Failed to commit');
-    res.status(400).json({ error: (err as Error).message ?? 'Unknown error' });
-  }
-}
-
-/** POST /api/git/stage */
-async function handleGitStage(ctx: RouteContext, req: Request, res: Response): Promise<void> {
-  const parsed = StageRequestSchema.safeParse(req.body);
-  if (!parsed.success) {
-    res.status(400).json({ error: 'chatId and files are required' });
-    return;
-  }
-
-  const { chatId, files } = parsed.data;
-  const workDir = ctx.chats.getEffectivePath(chatId);
-  if (!workDir) {
-    res.status(404).json({ error: 'Chat not found' });
-    return;
-  }
-
-  if (files.length === 0) {
-    res.json({ success: true });
-    return;
-  }
-
-  for (const file of files) {
-    const resolved = resolveAndValidatePath(workDir, file);
-    if (!resolved) {
-      res.status(400).json({ error: `Path outside project: ${file}` });
-      return;
-    }
-  }
-
-  try {
-    const svc = GitService.forProject(workDir);
-    await svc.stage(files);
-    res.json({ success: true });
-  } catch (err) {
-    logger.error({ err, workDir, chatId }, 'Failed to stage files');
-    res.status(400).json({ error: (err as Error).message ?? 'Unknown error' });
-  }
-}
-
-/** POST /api/git/unstage */
-async function handleGitUnstage(ctx: RouteContext, req: Request, res: Response): Promise<void> {
-  const parsed = StageRequestSchema.safeParse(req.body);
-  if (!parsed.success) {
-    res.status(400).json({ error: 'chatId and files are required' });
-    return;
-  }
-
-  const { chatId, files } = parsed.data;
-  const workDir = ctx.chats.getEffectivePath(chatId);
-  if (!workDir) {
-    res.status(404).json({ error: 'Chat not found' });
-    return;
-  }
-
-  if (files.length === 0) {
-    res.json({ success: true });
-    return;
-  }
-
-  for (const file of files) {
-    const resolved = resolveAndValidatePath(workDir, file);
-    if (!resolved) {
-      res.status(400).json({ error: `Path outside project: ${file}` });
-      return;
-    }
-  }
-
-  try {
-    const svc = GitService.forProject(workDir);
-    await svc.unstage(files);
-    res.json({ success: true });
-  } catch (err) {
-    logger.error({ err, workDir, chatId }, 'Failed to unstage files');
-    res.status(400).json({ error: (err as Error).message ?? 'Unknown error' });
-  }
-}
-
-/** POST /api/git/status */
-async function handleChatGitStatus(ctx: RouteContext, req: Request, res: Response): Promise<void> {
-  const parsed = StatusRequestSchema.safeParse(req.body);
-  if (!parsed.success) {
-    res.status(400).json({ error: 'chatId is required' });
-    return;
-  }
-
-  const { chatId } = parsed.data;
-  const workDir = ctx.chats.getEffectivePath(chatId);
-  if (!workDir) {
-    res.status(404).json({ error: 'Chat not found' });
-    return;
-  }
-
-  try {
-    const svc = GitService.forProject(workDir);
-    const output = await svc.statusRaw();
-    const { staged, unstaged, untracked } = parsePortcelainStatus(output);
-    res.json({ staged, unstaged, untracked });
-  } catch (err) {
-    if (!isNotGitRepo(err)) {
-      logger.warn({ err, workDir, chatId }, 'Failed to get git status for chat');
-    }
-    res.status(400).json({ error: (err as Error).message ?? 'Unknown error' });
-  }
-}
-
-/** POST /api/projects/:id/git/diff-since-main */
-async function handleDiffSinceMain(ctx: RouteContext, req: Request, res: Response): Promise<void> {
-  const chatId = req.body.chatId as string | undefined;
-  const files = Array.isArray(req.body.files) ? (req.body.files as string[]) : undefined;
-
-  const basePath = getEffectivePath(ctx, param(req, 'id'), chatId);
-  if (!basePath) {
-    res.status(404).json({ error: 'Project not found' });
-    return;
-  }
-
-  try {
-    const svc = GitService.forProject(basePath);
-
-    let baseBranch: string | null = null;
-    let mergeBase: string | null = null;
-    for (const base of ['main', 'master']) {
-      const sha = await svc.mergeBase(base, 'HEAD');
-      if (sha) {
-        baseBranch = base;
-        mergeBase = sha;
-        break;
-      }
-    }
-
-    if (!mergeBase) {
-      res.json({ diffs: {}, baseBranch: null, mergeBase: null });
-      return;
-    }
-
-    const nameStatusArgs = ['--name-status', `${mergeBase}..HEAD`, ...(files ? ['--', ...files] : [])];
-    const nameStatusOutput = await svc.diff(nameStatusArgs);
-    const changedFiles = parseDiffNameStatus(nameStatusOutput);
-
-    const diffs: Record<string, { main: string; worktree: string }> = {};
-    await Promise.all(
-      changedFiles.map(async ({ status, path, oldPath }) => {
-        let main = '';
-        let worktree = '';
-
-        if (!status.startsWith('A')) {
-          try {
-            main = await svc.show(`${mergeBase}:${oldPath ?? path}`);
-          } catch {
-            /* new file or binary */
-          }
-        }
-
-        if (!status.startsWith('D')) {
-          const resolvedPath = resolveAndValidatePath(basePath, path);
-          if (resolvedPath) {
-            try {
-              worktree = await readFile(resolvedPath, 'utf-8');
-            } catch {
-              /* deleted file */
-            }
-          }
-        }
-
-        diffs[path] = { main, worktree };
-      }),
-    );
-
-    res.json({ diffs, baseBranch, mergeBase });
-  } catch (err) {
-    logger.error({ err, basePath, chatId }, 'Failed to get diff since main');
-    res.status(400).json({ error: (err as Error).message ?? 'Unknown error' });
-  }
-}
-
 export function gitRoutes(ctx: RouteContext): Router {
   const router = Router();
 
-  router.post(
-    '/api/git/status',
-    asyncHandler((req, res) => handleChatGitStatus(ctx, req, res)),
-  );
-  router.post(
-    '/api/git/stage',
-    asyncHandler((req, res) => handleGitStage(ctx, req, res)),
-  );
-  router.post(
-    '/api/git/unstage',
-    asyncHandler((req, res) => handleGitUnstage(ctx, req, res)),
-  );
-  router.post(
-    '/api/git/commit',
-    asyncHandler((req, res) => handleGitCommit(ctx, req, res)),
-  );
-  router.post(
-    '/api/git/push',
-    asyncHandler((req, res) => handleGitPush(ctx, req, res)),
-  );
   router.get(
     '/api/projects/:id/git/branch-diffs',
     asyncHandler((req, res) => handleBranchDiffs(ctx, req, res)),
@@ -502,11 +202,8 @@ export function gitRoutes(ctx: RouteContext): Router {
     '/api/projects/:id/git/diff',
     asyncHandler((req, res) => handleDiff(ctx, req, res)),
   );
-  router.post(
-    '/api/projects/:id/git/diff-since-main',
-    asyncHandler((req, res) => handleDiffSinceMain(ctx, req, res)),
-  );
 
+  router.use(gitChatRoutes(ctx));
   router.use(gitWriteRoutes(ctx));
 
   return router;
